refactor(feedback): document-flow page on mobile, dialog on desktop
The feedback form has a textarea and a submit button, which a fixed overlay
can't keep above the iOS keyboard (same ceiling as search). On mobile the
trigger now navigates to /feedback — a document-flow page where the whole page
scrolls, so the submit button is always reachable. Desktop has no keyboard
constraint and keeps the centered dialog.

- Extract the form body into FeedbackForm (shared by dialog + page); the action
  row lives in the scroll stream, not a sticky footer.
- /feedback (noindex) reconstructs the lens context and, for data_issue, the
  reportable-field list server-side from ?mount&lensId via resolveLensSpecAndFields
  — nothing rich crosses the URL, and the field list matches the detail page by
  construction. ?q carries a general feedback search query.
- FeedbackTrigger branches on useBreakpoint: navigate on mobile, open dialog on
  desktop. All ~10 trigger sites flow through it unchanged.
